web_search/project/openai_client.py
from openai import AsyncOpenAI
import os
import httpx

SGP_API_KEY = os.getenv("SGP_API_KEY")
SGP_BASE_URL = os.getenv("SGP_BASE_URL")
SGP_ACCOUNT_ID = os.getenv("SGP_ACCOUNT_ID")
OPENAI_ORG_ID = os.getenv("OPENAI_ORG_ID", None)
max_retries = os.getenv("OPENAI_MAX_RETRIES", 100)
timeout = os.getenv("OPENAI_TIMEOUT", 10)

MAX_TOKENS = os.environ.get("OAI_MAX_TOKENS", 10000)
OAI_MODEL = os.environ.get("OAI_MODEL", "openai/gpt-4o")
TEMPERATURE_B100 = int(os.environ.get("OAI_TEMPERATURE_B100", 0))

http_client = httpx.AsyncClient(verify=False)

openai_client = AsyncOpenAI(
            base_url=SGP_BASE_URL,
            api_key="",
            default_headers = {
              "x-api-key": SGP_API_KEY,
              "x-selected-account-id": SGP_ACCOUNT_ID
            },
            http_client=http_client,
        )
