web_search/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web_search"
version = "0.1.0"
description = "An AgentEx agent"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "agentex-sdk<0.6.0",
    "scale-gp",
    "openai>=1.0.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.hatch.build.targets.wheel]
packages = ["project"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88 
